split/requirements.txt
Django==1.11.5
gunicorn==19.7
django-localflavor==1.5.3
